requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
discord.py==2.2.3
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
mediapipe==0.10.0
multidict==6.0.4
numpy==1.24.3
opencv-contrib-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sounddevice==0.4.6
yarl==1.9.2
